Alter dkan_datasets node_options via strongarm
Dkan_datasets exports the node_options_[dataset|resource] variables which
contains containes the publishing options for the content type. Workbench module
override this variable for it's own use which leads to a conflict and the
dkan_workflow settings being lost after each feature revert of dkan_dataset.

Since using features_overrides in dkan modules is not allowed. We had to use the
hook_strongarms_alter to add the 'revision' and 'moderation' values to the
node_options_[dataset|resource] variables.